Enable AWS STS 'assume role' credentials (#23)
Enable AWS STS temporary 'assume role' credentials so that we don't need to store AWS secret keys.

Update merge_vars.py to more recent plugin conventions and add functionality to merge inline variables ('facts' option), as well as files. This is needed for the STS logic because we need to update cluster_vars with the temporary STS credentials and we cannot use set_fact, because set_fact causes the source of the variables to change to a higher precedence (facts are higher than include_vars), and if you do this, then merge_vars will not be able to override the values later. Also, fix merge_vars so that it does not remove all the previously-defined ansible_facts, and add the ability to merge in the new facts either replacing previously-defined, or merging recursively ('replace' option).

Also - save memory by not doing gather_facts for clusterverse_confg, instead, only get the subset of facts needed.

* Update scaleup/scaledown README.md and Jenkinsfile_testsuite to new logic used to combine cluster_vars with cluster_vars_override

* Simplify vars combining